add possibility to remove account and logout

+++ src/main/java/com/wardyn/Projekt2/controllers/web/UserWebController.java
@@ -14,6 +14,8 @@ import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.List;
@@ -191,13 +193,32 @@ public class UserWebController {
     }
 
     @DeleteMapping("/users/delete/{id}")
-    public String deleteUser(@PathVariable Long id, RedirectAttributes redirectAttributes) {
-        Boolean deleted = userService.deleteUser(id);
-        if (deleted.equals(false)) {
-            redirectAttributes.addFlashAttribute("error", "There is no user with given id to delete");
-            return "redirect:/users";
+    public String deleteUser(@PathVariable Long id, RedirectAttributes redirectAttributes, @CookieValue(value = "id", defaultValue = "-1") String cookieId, HttpServletResponse response) {
+        Long parsedId = Long.parseLong(cookieId);
+
+        Optional<User> loggedUser = userService.getUserById(parsedId);
+
+        if (loggedUser.isPresent()) {
+            User user = loggedUser.get();
+            if (user.getRole().equals(Role.ADMIN) || parsedId.equals(id)) {
+                Boolean deleted = userService.deleteUser(id);
+
+                if (deleted.equals(false)) {
+                    redirectAttributes.addFlashAttribute("error", "There is no user with given id to delete");
+                    return "redirect:/users";
+                }
+
+                if (id.equals(Long.parseLong(cookieId))) {
+                    Cookie cookie = new Cookie("id", "-1");
+                    response.addCookie(cookie);
+                    return "redirect:/";
+                }
+                
+                return "redirect:/users";
+            }
         }
-        return "redirect:/users";
+
+        return "redirect:/";
     }
 
     private List<List<Object>> createListOfApps() {
